+++ include/xtd/math/asinh.h
@@ -21,7 +21,7 @@ namespace xtd {
     return __float2half(::asinhf(__half2float(arg)));
 #elif defined(XTD_TARGET_HIP)
     // HIP/ROCm device code
-    return __float2half_rn(::asinhf(__half2float(arg)));
+    return __float2half(::asinhf(__half2float(arg)));
 #elif defined(XTD_TARGET_SYCL)
     // SYCL device code
     return sycl::asinh(static_cast<sycl::half>(arg));
