+++ include/xtd/math/sinh.h
@@ -18,12 +18,11 @@ namespace xtd {
   XTD_DEVICE_FUNCTION inline constexpr float16 sinh(float16 arg) {
 #if defined(XTD_TARGET_CUDA)
     // CUDA device code
-    return ::sinhf(static_cast<float>(arg));
+    return __float2half(::sinhf(__half2float(arg)));
 #elif defined(XTD_TARGET_HIP)
     // HIP/ROCm device code
-    // return ::hcosh(arg);
     // AMD uses a suboptimal range reduction, use full float evaluation
-    return ::sinhf(static_cast<float>(arg));
+    return __float2half(::sinhf(__half2float(arg)));
 #elif defined(XTD_TARGET_SYCL)
     // SYCL device code
     return sycl::sinh(static_cast<float>(arg));
